updated script names
