+++ scripts/application-detection.gmp.py
@@ -16,42 +16,44 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import time
-
 
 def check_args(args):
     len_args = len(args.script) - 1
-
     if len_args is not 1:
         message = """
-        This script deletes overrides with a specific filter value
+        This script will display all hosts with the searched applications!
 
-        <filter>  -- the parameter for the filter.
+        1. <application>  -- Name of the application
 
         Example:
             $ gvm-script --gmp-username name --gmp-password pass \
-    ssh --hostname <gsm> scripts/delete-overrides-by-filter.gmp <filter>
+    ssh --hostname <gsm> scripts/application-detection.gmp.py <application>
         """
         print(message)
         quit()
 
 
-def delete_overrides(gmp, filter_value):
-    filters = gmp.get_overrides(filter=filter_value)
+def print_assets(gmp, appname):
+    res = gmp.get_reports()
 
-    if not filters.xpath('override'):
-        print('No overrides with filter: %s' % filter_value)
+    hosts = res.xpath('//host')
 
-    for f_id in filters.xpath('override/@id'):
-        print('Delete override: %s' % f_id, end='')
-        res = gmp.delete_override(f_id)
-
-        if 'OK' in res.xpath('@status_text')[0]:
-            print(' OK')
+    for host in hosts:
+        ip = host.xpath('ip/text()')
+        hostname = host.xpath('detail/name[text()="hostname"]/../value/text()')
+        if len(hostname) is 0:
+            hostname = ""
         else:
-            print(' ERROR')
+            hostname = hostname[0]
 
-        time.sleep(60)
+        print('{ip} ({hostname})'.format(ip=ip, hostname=hostname))
+        apps = host.xpath(
+            'detail/name[text() = "App"]/../value['
+            'contains(text(), "{0}")]/text()'.format(appname)
+        )
+        for app in apps:
+            print('\t' + app)
+        print('\n')
 
 
 def main(gmp, args):
@@ -59,9 +61,7 @@ def main(gmp, args):
 
     check_args(args)
 
-    filter_value = args.script[1]
-
-    delete_overrides(gmp, filter_value)
+    print_assets(gmp, args.script[1])
 
 
 if __name__ == '__gmp__':
