+++ scripts/clean-sensor.gmp.py
