+++ scripts/send-tasks.gmp.py
@@ -32,7 +32,7 @@ def check_args(args):
 
         Example:
             $ gvm-script --gmp-username name --gmp-password pass \
-    ssh --hostname <gsm> scripts/send-tasks.gmp example_file.xml
+    ssh --hostname <gsm> scripts/send-tasks.gmp.py example_file.xml
         """
 
         print(message)
