+++ scripts/update-task-target.gmp.py
@@ -32,7 +32,7 @@ def check_args(args):
 
         Example for starting up the routine:
             $ gvm-script --gmp-username name --gmp-password pass \
-    ssh --hostname <gsm> scripts/update-task-target-hosts.gmp hosts_file.csv \
+    ssh --hostname <gsm> scripts/update-task-target.gmp.py hosts_file.csv \
     "303fa0a6-aa9b-43c4-bac0-66ae0b2d1698"
 
         """
